LayoutTests/fast/dom/length-property-configurable-expected.txt

The 'length' property on interfaces objects should have attributes { [[Writable]]: false, [[Enumerable]]: false, [[Configurable]]: true }.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


* Event
PASS descriptor.writable is false
PASS descriptor.enumerable is false
PASS descriptor.configurable is true
PASS descriptor.value is 1
* Element
PASS descriptor.writable is false
PASS descriptor.enumerable is false
PASS descriptor.configurable is true
PASS descriptor.value is 0
* Document
PASS descriptor.writable is false
PASS descriptor.enumerable is false
PASS descriptor.configurable is true
PASS descriptor.value is 0
* Window
PASS descriptor.writable is false
PASS descriptor.enumerable is false
PASS descriptor.configurable is true
PASS descriptor.value is 0
* HTMLOptionElement
PASS descriptor.writable is false
PASS descriptor.enumerable is false
PASS descriptor.configurable is true
PASS descriptor.value is 0
* Option
PASS descriptor.writable is false
PASS descriptor.enumerable is false
PASS descriptor.configurable is true
PASS descriptor.value is 0
PASS successfullyParsed is true

TEST COMPLETE
